+++ dashboard/templates/dashboard/add_character.html
@@ -1,14 +1,12 @@
-{% extends 'core/base.html' %}
+{% extends 'core/base.html'%}
 
 {% block title %}Add Character{% endblock %}
 
 {% block content %}
-<div class="w3-content">
-  <h1>Add Character</h1>
-  <form method="POST">
+<h1>Add Character</h1>
+<form method="POST">
     {% csrf_token %}
     {{ form }}
     <input class="w3-button w3-teal w3-round" type="submit" value="Add"/>
-  </form>
-</div>
+</form>
 {% endblock %}
