+++ dashboard/urls.py
@@ -5,4 +5,6 @@ app_name = 'dashboard'
 
 urlpatterns = [
     path('', views.dashboard, name='dashboard'),
+    path('add-character', views.add_character, name='add-character'),
+    path('add-light-cone', views.add_light_cone, name='add-light-cone'),
 ]
